Move things into their own modules and add picker combinators
